ARM: OMAP4: hwmod_data: Remove OMAP4 DSP hwmod data
commit 9544563496732a80b38106355851adbc590d2619 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The OMAP4 DSP device now fully relies on the omap-prm reset
and ti-sysc drivers for managing the resets and clocks. The
dts node has been updated to remove the ti,hwmods reference
as well. So drop the legacy hwmod data for the device.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>