dmaengine: ti: k3-udma: Use the TR counter helper for slave_sg and cyclic
commit 691628b4affd8864282833671b65ad726dc9318d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Use the generic TR setup function to get the TR counters for both cyclic
and slave_sg transfers.
This way the period_size for cyclic and sg_dma_len() for slave_sg can be
as large as (SZ_64K - 1) * (SZ_64K - 1) and we can handle cases when the
length is >SZ_64K and a prime number.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>