dmaengine: ti: k3-udma: Fix terminated transfer handling
commit 8ba314ce86b66762c6ff1effff4641cc57b3d53b from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When we receive back the descriptor of the terminated transfer the cookie
must be marked as completed to make sure that the accounting is correct.

In udma_tx_status() the status should be marked as completed if the channel
is no longer running (it can only happen if the channel is not yet started
for the first time, or after a channel termination).

Fixes: 25dcb5dd7b7ce ("dmaengine: ti: New driver for K3 UDMA")
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>