dt-bindings: remoteproc: omap: Update ti,autosuspend-delay-ms description
commit 4111ff5f0d09c5f09bb0a762f314303eb1df5aa0 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The OMAP remoteproc binding has been updated to add some more details
for the "ti,autosuspend-delay-ms" property that is used for the default
runtime pm autosuspend delay in Linux. This is an optional property
with a default value of 10000 ms, but the runtime auto-suspend can be
disabled for a specific remote processor by default by using a negative
value.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>