ARM: dts: hackfix DT overlay and merge builds
commit b9c6ae1de9a9336d2e89e6156c34829e718d198d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The dependencies do not work for DT overlays and DT merge. As a hack
fix, always run the rules for DT overlays and merge.

As this causes the dtb-merge to be run for every build, even if nothing
changed, remove the print from dtb-merge script to hide the confusion.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Acked-by: Tero Kristo <t-kristo@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>