arm64: dts: ti: am654-base-board: fix clock node name
commit e1f1d162e4b8035b1858ee898361026c2d82dab1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The fixed clock for OV5640 is named 'clock' which is a very generic name
and easily leads to conflicts. I encountered this with a similarly named
fixed-clock node in k3-am654-evm-tc358876.dtso, which then overrode the
OV5640 fixed clock, causing OV5640 not to work when tc358876 overlay had
been loaded.

Rename the node to 'fixed-clock-ov5640'.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>