media: ti-vpe: cal: use runtime_resume for errata handling
commit e93a8f2f61671e7630c6813a65c538f25bd3bb54 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

We need to do errata handling every time CAL is being enabled. The code
is currently in cal_runtime_get(), which is not the correct place for
it.

Move the code to cal_runtime_resume, which is called every time CAL is
enabled.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Tested-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Reviewed-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>