clocksource/drivers/timer-ti-dm: Add support to get pwm current status
commit 5cd2d045a09473afda0b1a80ed5ab1cc88488480 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 92fd86864ec4ac089de47f0f1f4c47418b343448 upstream.

omap_dm_timer_ops provide support to configure the pwm but there is no
support to get the current status. For configuring pwm it is advised to
check the current hw status instead of relying on pwm framework. So
implement a new timer ops to get the current status of pwm.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Acked-by: Tony Lindgen <tony@atomide.com>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Link: https://lore.kernel.org/r/20200305082715.15861-6-lokeshvutla@ti.com
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>