clocksource/drivers/timer-ti-dm: Do one override clock parent in prepare()
commit 939187d0fea497e62e5322a795452ef3bf2d1ef0 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

omap_dm_timer_prepare() is setting up the parent 32KHz clock. This
prepare() gets called by request_timer in the client's driver. Because of
this, the timer clock parent that is set with assigned-clock-parent is being
overwritten. So drop this default setting of parent in prepare().

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Reviewed-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>