arm: dts: add am335x-boneblack-pps.dts
commit 95fb7e666cb683d41b9492f0a6c9613e890ea6cd from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add a new am335x-boneblack-pps.dts that can be used to configure TIMER7
in PWM mode. This PWM signal can be used as a PPS signal when
synchronized to PTP clock. Typically this synchronization is done by a
userspace program. This PPS signal can be observed on pin P8.8

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>