arm: dts: ti: dt-overlay: Add a pps overlay for am57xx idk platforms
commit fe7d00069a0239d07c51d1f3c4c86cf8a0945c9c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add a new am57xx-idk-pps.dtso that can be used to configure TIMER16
in PWM mode. This PWM signal can be used as a PPS signal when
synchronized to PTP clock. Typically this synchronization is done by a
userspace program.

Set the pwm parent as abe clk that is configured to 100MHz. This gives
a minimum sampling error of just 10ns and provide a better
synchronization technique.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>