arm64: dts: ti: k3-j721e-main: Configure MAIN R5FSS1 for Split-mode
commit 746e869b8cebe091d27b8d7b95a613d5bcae59f0 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Switch the MAIN R5FSS1 cluster to be configured for Split-mode as the
default so that two different applications can be run on each of the
R5F cores in performance mode. LockStep-mode would be available only
on SoCs efused with the appropriate capability bit, and Split-mode is
the mode that is available on all J721E SoCs.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>