gpio-siul: Allocate 'irq_chip' data structure per instance
commit 2dc012a4fa42e1ae5f9bee91193e7aeecfed6a86 from
https://source.codeaurora.org/external/autobsps32/linux

In Linux Kernel 5.4, a warning message is thrown on Kernel boot,
during the probe of the second and later instances of the gpio-siul
chips.

The issue is that the driver is using the same 'siul2_gpio_irq_chip'
data structure between multiple instances.

We fix this by allocating the 'irq_chip' data structure per instance.

Issue: ALB-4883
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>