arm64: dts: k3-am654-base-board-sr1: Disable UHS modes in SR1.0
commit 0d6d7c1309bbb8b9006a3bb007f707b268585543 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The 4 bit interface in am65 sr1.0 cannot be supported at 3.3V or 1.8V
because of erratas i2025 and i2026. However, the SD card is the primary
boot mode for development usecases.

Continue to enable SD card in SR1.0 and disable UHS modes in it to
minimise any ageing issues happening because of the erratas.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>