net: hsr/prp: add support for mc filtering at the slave device
commit 6ce95c1bcf7d6882994080e8b2e0b48cd57f77f6 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When MC list is updated by the networking layer due to a user
command and as well as when allmulti flag is set, it needs to be
passed to the enslaved Ethernet devices. This patch allows this
to happen by implementing ndo_change_rx_flags() and ndo_set_rx_mode()
API calls that in turns pass it to the slave devices using
existing API calls.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Reviewed-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>