TEMP: ARM: dts: am335x: Add PRUSS IEP clock mux
commit 961e201faa15b1313d6e6038c80c5def21724992 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The PRUSS module has an internal clock mux for the IEP source clock,
that can be sourced from either of the PRU_ICSS_IEP_GCLK or the
PRU_ICSS_OCP_GCLK and configured through a register in the PRUSS
CFG sub-module. Add a clock mux node for this. The default source
for this mux clock is PRU_ICSS_IEP_GCLK.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>