TEMP: ARM: dts: am4372: Add PRUSS IEP clock muxes
commit 44188ba7df74fe0ddd933b9382bbe550888d81b6 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The PRUSS module has an internal clock mux for the IEP source clock,
that can be sourced from either of the PRU_ICSS_IEP_GCLK or the
PRU_ICSS_OCP_GCLK and configured through a register in the PRUSS
CFG sub-module. Add a clock mux node for this in each of the PRUSS
instances on AM437x SoCs. The default source for these mux clocks
is the PRU_ICSS_IEP_GCLK clock.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>