TEMP: ARM: dts: keystone-k2g: Add PRUSS IEP clock muxes
commit 5ba8b9d4c3cd24ab61e06a4c424e4c4e1aee3d4f from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The PRUSS module has an internal clock mux for the IEP source clock,
that can be sourced from either of the ICSS_IEP_CLK or the ICSS_VCLK_CLK
and configured through a register in the PRUSS CFG sub-module. Add a
clock mux node for this in each of the PRUSS instances on 66AK2G SoCs.
The default source for these mux clocks is the ICSS_IEP_CLK clock.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>