mtd: spi-nor: perform a Soft Reset on shutdown
commit 52d28830370d3f92615af07f4f5f177c7e0d1bc6 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Perform a Soft Reset on shutdown on flashes that support it so that the
flash can be reset to its initial state and any configurations made by
spi-nor (given that they're only done in volatile registers) will be
reset. This will hand back the flash in pristine state for any further
operations on it.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>