mtd: spi-nor: allow using MT35XU512ABA in Octal DTR mode
commit ef46ba7c0addfcdd365579ca9c0a1458bccaaf35 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Since this flash doesn't have a Profile 1.0 table, the Octal DTR
capabilities are enabled in the post SFDP fixup, along with the 8D-8D-8D
fast read settings.

Enable Octal DTR mode with 20 dummy cycles to allow running at the
maximum supported frequency of 200Mhz.

The flash supports the soft reset sequence. So, add the flag in the
flash's info.

Add a forward definition for spi_nor_default_setup(). It is not moved to
the top because it sits close to all its related functions right now and
the new position would be a bit awkward.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>