arm64: dts: s32: Switch to DSPI_TCFQ_MODE
commit d3052f70cd3f6c48b7929e1854994ce78b7f444f from
https://source.codeaurora.org/external/autobsps32/linux

The only advantage of DSPI_EOQ_MODE is a reduced number of interrupts,
of one per queue (where queues are the same size as the TX and RX
FIFOs) instead of one per frame. This also avoids some inter-frame
delays. However, it comes with the drawback of the CS being deasserted
after each queue, which limits the maximum transaction size to 5 frames,
as received by slave devices.

Therefore, we switch to DSPI_TCFQ_MODE for all S32 SoCs.

Issue: ALB-4486
Signed-off-by: Stefan-Gabriel Mirea <stefan-gabriel.mirea@nxp.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>