dmaengine: fsl-edma: Store interrupts on a per-instance basis
commit ec7023ecf2b11f7ee47172c3ed4e1829c887c548 from
https://source.codeaurora.org/external/autobsps32/linux

The eDMA driver stores the IRQ numbers in the elements of the
platform-specific *_edma_irqs arrays. When there are more eDMA instances
(as on S32 GEN1), fsl_edma_irq_init initializes the irqno fields with
the interrupt lines found in the dts node of the first eDMA, and then
overwrites them at subsequent calls when the other eDMAs are probed.
Therefore, fsl_edma_irq_exit can only free the IRQs of the last instance
probed.

Issue: ALB-4692
Signed-off-by: Stefan-Gabriel Mirea <stefan-gabriel.mirea@nxp.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>