net: ti: prueth: hsr: Add support for timestamping tx packets
commit 3b9b5f3413411b7085e8edde0b623c19cf45f84b from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When PTP support is enabled in hsr firmware, the firmware timestamps
the ptp packets that are going out via prueth ports. The timestamp value
is copied to a shared memory location. After timestamping, a host irq is
raised by the firmware as a notification to read the timestamp.

Add support for handling this ptp tx irq, read the tx timestamp and pass
it to userspace layer.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>