remoteproc/k3-r5: Fix TCM initialization upon reset deassertion failures
commit f8dd7dc1c07c50e04bd6e20dc7069cc098f7ee93 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The R5F TCMs are zeroed out for proper ECC functionality in the .prepare()
callback, and is performed after the required resets are deasserted. This
logic however is incorrectly ignoring the status of the TI-SCI proc
control APIs for releasing the resets, and would result in a bus error.
Fix this properly for the reset deassertion failure path.

Fixes: 002de8948d39 ("remoteproc/k3-r5: Initialize TCM memories for ECC")
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>