usb: cdns3: Enable workaround for USB2.0 PHY Rx compliance test PHY lockup
commit 2c2e094858f2f8f21afba3a2f5d67a56337565bd from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

USB2.0 PHY hangs in Rx Compliance test when the incoming packet
amplitude is varied below and above the Squelch Level of
Receiver during the active packet multiple times.

Version 1 of the controller allows PHY to be reset when RX fail condition
is detected to work around the above issue. This feature is
disabled by default and needs to be enabled using a bit from
the newly added PHYRST_CFG register. This patch enables the workaround.

Signed-off-by: Pawel Laszczak <pawell@cadence.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>