Extended the TMU driver so that both gen1 and s32v234 are supported.
commit 7580df009dad57cef7fb520424d3d903a3428b90 from
https://source.codeaurora.org/external/autobsps32/linux

This driver uses up to 3 sites of the TMU to measure temperature.
Word of warning: there is a HW bug with the sensor on S32G cut1.
This bug results in noisy/unreliable temperature measurements.
The calibration table used here was determined experimentally
and should be replaced on cut2. One can use the _sensors_ command
from the lm-sensors package to get the temperature provided by
the TMU.

Signed-off-by: Dorin Ionita <dorin.ionita@nxp.com>
Issue: ALB-5142
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>