dmaengine: Fix return value for dma_request_chan() in case of failure
commit e5f5163f856285b9ca35597275de48c3b5e6dcb5 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 474809a28e7b0671a5090de6e0db91f0e3821360 upstream.

Commit 71723a96b8b1 ("dmaengine: Create symlinks between DMA channels and
slaves") changed the dma_request_chan() function flow in such a way that
it always returns EPROBE_DEFER in case of channels that cannot be found.
This break the operation of the devices which have optional DMA channels
as it puts their drivers in endless deferred probe loop. Fix this by
propagating the proper error value.

Fixes: 71723a96b8b1 ("dmaengine: Create symlinks between DMA channels and slaves")
Signed-off-by: Marek Szyprowski <m.szyprowski@samsung.com>
Reviewed-by: Geert Uytterhoeven <geert+renesas@glider.be>
Link: https://lore.kernel.org/r/20200130070834.17537-1-m.szyprowski@samsung.com
[vkoul: fix typo in patch title]
Signed-off-by: Vinod Koul <vkoul@kernel.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>