dmaengine: Cleanups for the slave <-> channel symlink support
commit 47f96b43028262d827caec7dc419ae768d8efe59 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit bad83565eafe8a00922ad4eed6920625a10a2126 upstream.

No need to use goto to jump over the
return chan ? chan : ERR_PTR(-EPROBE_DEFER);
We can just revert the check and return right there.

Do not fail the channel request if the chan->name allocation fails, but
print a warning about it.

Change the dev_err to dev_warn if sysfs_create_link() fails as it is not
fatal.

Only attempt to remove the DMA_SLAVE_NAME symlink if it is created - or it
was attempted to be created.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Reviewed-by: Geert Uytterhoeven <geert+renesas@glider.be>
Link: https://lore.kernel.org/r/20200131093859.3311-2-peter.ujfalusi@ti.com
Signed-off-by: Vinod Koul <vkoul@kernel.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>