remoteproc/k3-dsp: Remove unneeded pm_runtime usage
commit acd58460dab1ac91264a7a50ea75921a5d62cf15 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The TI K3 DSP remoteproc driver uses couple of pm_runtime API in probe
and remove to turn on and off the necessary clocks associated with the
DSP internal memories. However, these are no-ops for this driver, as
the DTS nodes do not use any power-domains, and so there is no real
back-end behind these calls. These clocks are instead turned on and
off during the .prepare() and .unprepare() callbacks. So, eliminate
the dead pm_runtime API usage from the driver.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>