fsl_linflexuart: Resume DMA transfer after polling mode finished
commit 2ab64df542e9d47263df4a62812066b14529d6c5 from
https://source.codeaurora.org/external/autobsps32/linux

When there are both kernel and userspace output, the
last one can't be seen after a while. The userspace console hangs until
the user wakes up it by sending a character.
The fix is to resume the tty output and the DMA transfer after polling
mode output is finished.

Issue: ALB-5329
Signed-off-by: Js Ha <js.ha@nxp.com>
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>