remoteproc/k3-r5: Fix initialization logic in k3_r5_rproc_configure()
commit c308fd8c0c1c3f258021d9e9518ef6f5af68c8a8 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The function k3_r5_rproc_configure() is used to initialize various R5F
cluster and core level configurations to some sane default values during
the probe function. One of the steps it does is to program the individual
cores symmetrically in LockStep-mode and does this by initially configuring
the cluster to Split-mode and then switching back the mode once all the
core configurations are done. This logic used a wrong variable for
halting both the cores. Fix this.

This clever logic is confusing for developers and reviewers alike, so
improve the code readability by adding some detailed function description.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>